+++ src/constant.ts
@@ -2,9 +2,9 @@ import type { StatInkPostBody, VsHistoryDetail } from "./types.ts";
 
 export const AGENT_NAME = "splashcat / s3si.ts";
 export const AGENT_VERSION = "1.1.1";
-export const S3SI_VERSION = "0.4.1";
+export const S3SI_VERSION = "0.4.4";
 export const COMBINED_VERSION = `${AGENT_VERSION}/${S3SI_VERSION}`;
-export const NSOAPP_VERSION = "2.5.1";
+export const NSOAPP_VERSION = "2.5.2";
 export const WEB_VIEW_VERSION = "4.0.0-d5178440";
 export enum Queries {
   HomeQuery = "7dcc64ea27a08e70919893a0d3f70871",
@@ -12,6 +12,7 @@ export enum Queries {
   RegularBattleHistoriesQuery = "3baef04b095ad8975ea679d722bc17de",
   BankaraBattleHistoriesQuery = "0438ea6978ae8bd77c5d1250f4f84803",
   XBattleHistoriesQuery = "6796e3cd5dc3ebd51864dc709d899fc5",
+  EventBattleHistoriesQuery = "9744fcf676441873c7c8a51285b6aa4d",
   PrivateBattleHistoriesQuery = "8e5ae78b194264a6c230e262d069bd28",
   VsHistoryDetailQuery = "9ee0099fbe3d8db2a838a75cf42856dd",
   CoopHistoryQuery = "91b917becd2fa415890f5b47e15ffb15",